SL-20: update nav icons, reorganise pages
