--- src/components/Svg/svg.vue
@@ -1,17 +1,0 @@
-<template>
-  <svg class="my-svg-component">
-    <!-- Actual SVG paths, shapes, etc. here -->
-  </svg>
-</template>
-
-<script>
-export default {
-  name: 'MySVGComponent',
-};
-</script>
-
-<style scoped>
-.my-svg-component {
-  /* Scoped CSS here */
-}
-</style>
