+++ src/views/MessagePage.vue
@@ -1,5 +1,7 @@
 <template>
-  <div class="text-center">Message</div>
+  <div class="flex justify-center items-center h-screen text-white">
+    <h1>Welcome, Message</h1>
+  </div>
 </template>
 
 <script>
