+++ tailwind.config.js
@@ -8,38 +8,40 @@ export default {
       },
       colors: {
         midnight: '#000024',
+        green: {
+          light: '#9E9E9E',
+          DEFAULT: '#006400',
+          dark: '#006400',
+        },
         purple: {
           light: '#A3B3ED',
-          DEFAULT: '#6580E1',
+          default: '#6580E1',
           dark: '#455DB3',
         },
         pink: {
           light: '#E3D1EC',
-          DEFAULT: '#D0B2E0',
+          default: '#D0B2E0',
           dark: '#AC83C2',
         },
         main: {
           light: '#2D1577',
-          DEFAULT: '#140935',
+          default: '#140935',
           dark: '#0C0520',
         },
         card: {
           light: '#332E46',
-          DEFAULT: '#221E2F',
+          default: '#221E2F',
           dark: '#1B1826',
         },
         trans: {
           light: 'rgba(255,255,255,0.3)',
-          DEFAULT: 'rgba(255,255,255,0.2)',
-          dark: 'rgba(255,255,255,0.1)',
+          default: 'rgba(255,255,255,0.2)',
+          dark: '#FFFFFF1A',
         },
       },
       fontFamily: {
         sen: ['Sen', 'sans-serif'],
       },
-      // fontFamily: {
-      //   sen: ['Sen', 'sans-serif'],
-      // },
     },
   },
   plugins: [],
